nm.secrets: move get_secrets() to a common file
get_secrets() will be used from WireGuard support. Therefore, it should
be shared in order to do not duplicate code.
